Moved identity activation to it's own file.
