+++ src/activation/mod.rs
@@ -1,8 +1,10 @@
 use std::f64;
 
 pub mod sigmoid;
+pub mod identity;
 
 pub use self::sigmoid::Sigmoid;
+pub use self::identity::Identity;
 
 /// Activation functions
 pub trait Activation {
@@ -11,24 +13,6 @@ pub trait Activation {
     fn calc(&self, x: f64) -> f64;
     // Derivative
     fn derivative(&self, x: f64) -> f64;
-}
-
-pub struct Identity;
-
-impl Activation for Identity {
-    fn new() -> Identity {
-        return Identity;
-    }
-
-    /// Calculates the Identity of input `x`
-    fn calc(&self, x: f64) -> f64 {
-        x
-    }
-
-    /// Calculates the Derivative Identity of input `x`
-    fn derivative(&self, x: f64) -> f64 {
-        1f64
-    }
 }
 
 pub struct HyperbolicTangent;
@@ -97,22 +81,9 @@ impl Activation for RectifiedLinearUnit {
 #[cfg(test)]
 mod tests {
     use super::Activation;
-    use super::Identity;
     use super::HyperbolicTangent;
     use super::SoftPlus;
     use super::RectifiedLinearUnit;
-
-    #[test]
-    fn identity_test() {
-        let activation = Identity::new();
-        assert_approx_eq!(activation.calc(5f64), 5f64);
-    }
-
-    #[test]
-    fn identity_derivative_test() {
-        let activation = Identity::new();
-        assert_approx_eq!(activation.derivative(15f64), 1f64);
-    }
 
     #[test]
     fn tanh_test() {
